// ui/packages/components/src/common/gcsim/GraphComponents/colors.d.ts
export declare const Colors: {
	/** @deprecated use CERULEAN1 */
	COBALT1: string;
	/** @deprecated use CERULEAN2 */
	COBALT2: string;
	/** @deprecated use CERULEAN3 */
	COBALT3: string;
	/** @deprecated use CERULEAN4 */
	COBALT4: string;
	/** @deprecated use CERULEAN5 */
	COBALT5: string;
	CERULEAN1: string;
	CERULEAN2: string;
	CERULEAN3: string;
	CERULEAN4: string;
	CERULEAN5: string;
	FOREST1: string;
	FOREST2: string;
	FOREST3: string;
	FOREST4: string;
	FOREST5: string;
	GOLD1: string;
	GOLD2: string;
	GOLD3: string;
	GOLD4: string;
	GOLD5: string;
	INDIGO1: string;
	INDIGO2: string;
	INDIGO3: string;
	INDIGO4: string;
	INDIGO5: string;
	LIME1: string;
	LIME2: string;
	LIME3: string;
	LIME4: string;
	LIME5: string;
	ROSE1: string;
	ROSE2: string;
	ROSE3: string;
	ROSE4: string;
	ROSE5: string;
	SEPIA1: string;
	SEPIA2: string;
	SEPIA3: string;
	SEPIA4: string;
	SEPIA5: string;
	TURQUOISE1: string;
	TURQUOISE2: string;
	TURQUOISE3: string;
	TURQUOISE4: string;
	TURQUOISE5: string;
	VERMILION1: string;
	VERMILION2: string;
	VERMILION3: string;
	VERMILION4: string;
	VERMILION5: string;
	VIOLET1: string;
	VIOLET2: string;
	VIOLET3: string;
	VIOLET4: string;
	VIOLET5: string;
	BLUE1: string;
	BLUE2: string;
	BLUE3: string;
	BLUE4: string;
	BLUE5: string;
	GREEN1: string;
	GREEN2: string;
	GREEN3: string;
	GREEN4: string;
	GREEN5: string;
	ORANGE1: string;
	ORANGE2: string;
	ORANGE3: string;
	ORANGE4: string;
	ORANGE5: string;
	RED1: string;
	RED2: string;
	RED3: string;
	RED4: string;
	RED5: string;
	BLACK: string;
	DARK_GRAY1: string;
	DARK_GRAY2: string;
	DARK_GRAY3: string;
	DARK_GRAY4: string;
	DARK_GRAY5: string;
	GRAY1: string;
	GRAY2: string;
	GRAY3: string;
	GRAY4: string;
	GRAY5: string;
	LIGHT_GRAY1: string;
	LIGHT_GRAY2: string;
	LIGHT_GRAY3: string;
	LIGHT_GRAY4: string;
	LIGHT_GRAY5: string;
	WHITE: string;
};